00-3304: Show number of invited participants in data collection report

+++ src/main/java/org/olat/modules/quality/QualityService.java
@@ -129,7 +129,7 @@ public interface QualityService {
 	public List<QualityParticipation> loadParticipations(QualityDataCollectionLight dataCollection,
 			int firstResult, int maxResults, SortKey... orderBy);
 
-	public int getExecutorParticipationCount(QualityExecutorParticipationSearchParams searchParams);
+	public Long getExecutorParticipationCount(QualityExecutorParticipationSearchParams searchParams);
 
 	public List<QualityExecutorParticipation> loadExecutorParticipations(Translator translator,
 			QualityExecutorParticipationSearchParams searchParams, int firstResult, int maxResults, SortKey... orderBy);
