+++ src/main/java/org/olat/modules/quality/manager/QualityServiceImpl.java
@@ -423,7 +423,7 @@ public class QualityServiceImpl
 	}
 
 	@Override
-	public int getExecutorParticipationCount(QualityExecutorParticipationSearchParams searchParams) {
+	public Long getExecutorParticipationCount(QualityExecutorParticipationSearchParams searchParams) {
 		return participationDao.getExecutorParticipationCount(searchParams);
 	}
